Extract shelley `propWitVKeys` in a separate module
to increase cohesion and enable reuse
